requirements.txt
basketball-reference-scraper==2.0.0
beautifulsoup4==4.14.2
certifi==2025.10.5
charset-normalizer==3.4.4
contourpy==1.3.0
cycler==0.12.1
fonttools==4.60.1
html5lib==1.1
idna==3.11
importlib_resources==6.5.2
jupyter==1.1.1
kiwisolver==1.4.7
lxml==6.0.2
matplotlib==3.9.4
numpy==2.0.2
nbainjuries==1.0.0
packaging==25.0
pandas==2.3.3
pillow==11.3.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
six==1.17.0
tzdata==2025.2
urllib3==2.5.0
zipp==3.23.0

